+++ R/cranDownloadsPlotSmoother.R
@@ -5,7 +5,7 @@ addSmoother <- function(x, complete, current.wk, f, span, wk1, y.nm) {
       smooth.data <- stats::loess(complete[, y.nm] ~
         as.numeric(complete$date), span = span)
     } else if (any(dat$partial)) {
-      smooth.data <- smoothMonthData(x, complete, current.wk, f, span, wk1,
+      smooth.data <- smoothWeekData(x, complete, current.wk, f, span, wk1,
         y.nm)
     } else {
       smooth.data <- stats::loess(dat[, y.nm] ~ as.numeric(dat$date),
@@ -15,7 +15,7 @@ addSmoother <- function(x, complete, current.wk, f, span, wk1, y.nm) {
     if (any(dat$in.progress)) {
       smooth.data <- stats::lowess(complete$date, complete[, y.nm], f = f)
     } else if (any(dat$partial)) {
-      smooth.data <- smoothMonthData(x, current.wk, f, span, wk1)
+      smooth.data <- smoothWeekData(x, complete, current.wk, f, span, wk1, y.nm)
     } else {
       smooth.data <- stats::lowess(dat$date, dat[, y.nm], f = f)
     }
@@ -23,12 +23,14 @@ addSmoother <- function(x, complete, current.wk, f, span, wk1, y.nm) {
   if (nrow(dat) > 7) {
     x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
     lines(x.date, smooth.data$fitted, col = "blue", lwd = 1.25)
+    title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
   } else {
     lines(smooth.data, col = "blue", lwd = 1.25)
+    title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
   }
 }
 
-smoothMonthData <- function(x, complete, current.wk, f, span, wk1, y.nm) {
+smoothWeekData <- function(x, complete, current.wk, f, span, wk1, y.nm) {
   dat <- x$cranlogs.data
   wk1.start <- dat$date[1]
   if (weekdays(x$from) == "Sunday") {
@@ -38,7 +40,7 @@ smoothMonthData <- function(x, complete, current.wk, f, span, wk1, y.nm) {
     sel <- dat$partial & dat$date == wk1.start
     wk1.partial <- dat[sel, ]
     wk1.backdate <- wk1.partial
-    wk1.backdate$count <- sum(wk1$cranlogs.data$count)
+    wk1.backdate$count <- sum(wk1$count)
     wk1.backdate$cumulative <- wk1.backdate$count
   }
   tmp <- rbind(wk1.backdate, complete)
@@ -82,15 +84,18 @@ addSinglePlotSmoother <- function(x, complete, current.wk, f, span,
   if (nrow(dat) > 7) {
     x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
     lines(x.date, smooth.data$fitted, col = "blue", lwd = 1.25)
+    title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
   } else {
-    lines(smooth.data$date, smooth.data[, y.nm], col = "blue")
+    lines(smooth.data$x, smooth.data$y, col = "blue")
+    title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
   }
 }
 
 addMultiPlotSmoother <- function(i, x, complete, cbPalette, f, span,
-  statistic, vars, wk1.backdate, tmp) {
+  statistic, vars, wk1.backdate) {
 
   dat <- x$cranlogs.data
+  dat <- dat[dat$package == x$packages[i], ]
   current.wk <- dat[nrow(dat), ]
   last.obs.date <- x$last.obs.date
 
@@ -101,9 +106,11 @@ addMultiPlotSmoother <- function(i, x, complete, cbPalette, f, span,
         as.numeric(smooth.data$date), span = span)
       x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
       lines(x.date, smooth.data$fitted, col = cbPalette[i])
+      title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
     } else {
       lines(stats::lowess(smooth.data$date, smooth.data[, statistic], f = f),
         col = cbPalette[i])
+      title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
     }
   } else if (any(dat$partial)) {
     tmp <- rbind(wk1.backdate, complete)
@@ -115,19 +122,23 @@ addMultiPlotSmoother <- function(i, x, complete, cbPalette, f, span,
         as.numeric(tmp$date), span = span)
       x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
       lines(x.date, smooth.data$fitted, col = cbPalette[i])
+      title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
     } else {
       smooth.data <- stats::lowess(tmp$date, tmp[, statistic], f = f)
-      lines(smooth.data[, vars], col = cbPalette[i])
+      lines(smooth.data$x, smooth.data$y, col = cbPalette[i])
+      title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
     }
   } else {
     if (nrow(dat) > 7) {
-      smooth.data <- stats::loess(tmp[, statistic] ~
-        as.numeric(tmp$date), span = span)
+      smooth.data <- stats::loess(dat[, statistic] ~ as.numeric(dat$date), 
+        span = span)
       x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
       lines(x.date, smooth.data$fitted, col = cbPalette[i])
+      title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
     } else {
-      lines(stats::lowess(tmp[dat$package == x$packages[i], vars], f = f),
+      lines(stats::lowess(dat[dat$package == x$packages[i], vars], f = f),
         col = cbPalette[i])
+      title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
     }
   }
 }
@@ -141,27 +152,30 @@ addRPlotSmoother <- function(x, complete, f, span, pltfrm, pltfrm.col,
     invisible(lapply(seq_along(complete), function(i) {
       tmp <- complete[[i]]
       if (nrow(dat) > 7) {
-        smooth.data <- stats::loess(tmp[, statistic] ~
-          as.numeric(tmp$date), span = span)
+        smooth.data <- stats::loess(tmp[, statistic] ~ as.numeric(tmp$date), 
+          span = span)
         x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
         lines(x.date, smooth.data$fitted, lwd = 1.5, col = pltfrm.col[i])
+        title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
       } else {
         smooth.data <- stats::lowess(tmp$date, tmp[, statistic], f = f)
         lines(smooth.data, lwd = 1.5, col = pltfrm.col[i])
+        title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
       }
     }))
   } else if (any(dat$partial)) {
     invisible(lapply(seq_along(complete), function(i) {
       tmp <- rbind(wk1.backdate[i, ], complete[[i]])
       if (nrow(tmp) > 7) {
-        smooth.data <- stats::loess(tmp[, statistic] ~
-          as.numeric(tmp$date), span = span)
+        smooth.data <- stats::loess(tmp[, statistic] ~ as.numeric(tmp$date), 
+          span = span)
         x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
         lines(x.date, smooth.data$fitted, lwd = 1.5, col =  pltfrm.col[i])
+        title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
       } else {
         smooth.data <- stats::lowess(tmp$date, tmp[, statistic], f = f)
-        lines(smooth.data$x, smooth.data$fitted, lwd = 1.5,
-          col = pltfrm.col[i])
+        lines(smooth.data$x, smooth.data$fitted, lwd = 1.5, col = pltfrm.col[i])
+        title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
       }
     }))
   } else {
@@ -172,9 +186,11 @@ addRPlotSmoother <- function(x, complete, f, span, pltfrm, pltfrm.col,
           as.numeric(tmp$date), span = span)
         x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
         lines(x.date, smooth.data$fitted, lwd = 1.5, col = pltfrm.col[i])
+        title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
       } else {
         smooth.data <- stats::lowess(tmp$date, tmp[, statistic], f = f)
         lines(smooth.data, lwd = 1.5, col = pltfrm.col[i])
+        title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
       }
     }))
   }
@@ -190,9 +206,11 @@ addRTotPlotSmoother <- function(dat, complete, f, span, statistic,
         as.numeric(tmp$date), span = span)
       x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
       lines(x.date, smooth.data$fitted, col = "blue", lwd = 1.25)
+      title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
     } else {
-      lines(stats::lowess(tmp$date, tmp[, statistic], f = f),
-        col = "blue", lwd = 1.25)
+      lines(stats::lowess(tmp$date, tmp[, statistic], f = f), col = "blue", 
+        lwd = 1.25)
+      title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
     }
   } else if (any(dat$partial)) {
     tmp <- rbind(wk1.backdate, complete)
@@ -201,9 +219,11 @@ addRTotPlotSmoother <- function(dat, complete, f, span, statistic,
         as.numeric(tmp$date), span = span)
       x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
       lines(x.date, smooth.data$fitted, col = "blue", lwd = 1.25)
+      title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
     } else {
-      lines(stats::lowess(tmp$date, tmp[, statistic], f = f),
-        col = "blue", lwd = 1.25)
+      lines(stats::lowess(tmp$date, tmp[, statistic], f = f), col = "blue",
+        lwd = 1.25)
+      title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
     }
   } else {
     if (nrow(dat) > 7) {
@@ -211,9 +231,11 @@ addRTotPlotSmoother <- function(dat, complete, f, span, statistic,
         as.numeric(dat$date), span = span)
       x.date <- as.Date(smooth.data$x, origin = "1970-01-01")
       lines(x.date, smooth.data$fitted, col = "blue", lwd = 1.25)
+      title(sub = paste("loess span =", round(span, 2)), cex.sub = 0.9)
     } else {
-      lines(stats::lowess(dat$date, dat[, statistic], f = f),
-        col = "blue", lwd = 1.25)
+      lines(stats::lowess(dat$date, dat[, statistic], f = f), col = "blue", 
+        lwd = 1.25)
+      title(sub = paste("lowess f =", round(f, 2)), cex.sub = 0.9)
     }
   }
 }